Frontend user auth -- finished login. Start sign up form p1/p3
