+++ motify/frontend/src/components/SignUpForms/SignUpFormPageOne/index.jsx
@@ -1,0 +1,3 @@
+import SignUpFormPageOne from "./SignUpFormPageOne"
+
+export default SignUpFormPageOne;
